chore: 컴포넌트 구조 수정
